+++ provider/resources.go
@@ -193,6 +193,7 @@ func Provider() tfbridge.ProviderInfo {
 			"linode_lke_cluster":           {Tok: makeResource(mainMod, "LkeCluster")},
 			"linode_firewall":              {Tok: makeResource(mainMod, "Firewall")},
 			"linode_object_storage_object": {Tok: makeResource(mainMod, "ObjectStorageObject")},
+			"linode_vlan":                  {Tok: makeResource(mainMod, "Vlan")},
 		},
 		DataSources: map[string]*tfbridge.DataSourceInfo{
 			"linode_account": {Tok: makeDataSource(mainMod, "getAccount")},