+++ sdk/python/pulumi_linode/_inputs.py
@@ -35,6 +35,7 @@ __all__ = [
     'ObjectStorageBucketCertArgs',
     'ObjectStorageKeyBucketAccessArgs',
     'StackScriptUserDefinedFieldArgs',
+    'VlanAttachedLinodeArgs',
     'GetStackScriptUserDefinedFieldArgs',
 ]
 
@@ -1718,6 +1719,61 @@ class StackScriptUserDefinedFieldArgs:
     @one_of.setter
     def one_of(self, value: Optional[pulumi.Input[str]]):
         pulumi.set(self, "one_of", value)
+
+
+@pulumi.input_type
+class VlanAttachedLinodeArgs:
+    def __init__(__self__, *,
+                 id: Optional[pulumi.Input[int]] = None,
+                 ipv4_address: Optional[pulumi.Input[str]] = None,
+                 mac_address: Optional[pulumi.Input[str]] = None):
+        """
+        :param pulumi.Input[int] id: The ID of the Linode.
+        :param pulumi.Input[str] ipv4_address: The IPv4 address of the Linode.
+        :param pulumi.Input[str] mac_address: The mac address of the Linode.
+        """
+        if id is not None:
+            pulumi.set(__self__, "id", id)
+        if ipv4_address is not None:
+            pulumi.set(__self__, "ipv4_address", ipv4_address)
+        if mac_address is not None:
+            pulumi.set(__self__, "mac_address", mac_address)
+
+    @property
+    @pulumi.getter
+    def id(self) -> Optional[pulumi.Input[int]]:
+        """
+        The ID of the Linode.
+        """
+        return pulumi.get(self, "id")
+
+    @id.setter
+    def id(self, value: Optional[pulumi.Input[int]]):
+        pulumi.set(self, "id", value)
+
+    @property
+    @pulumi.getter(name="ipv4Address")
+    def ipv4_address(self) -> Optional[pulumi.Input[str]]:
+        """
+        The IPv4 address of the Linode.
+        """
+        return pulumi.get(self, "ipv4_address")
+
+    @ipv4_address.setter
+    def ipv4_address(self, value: Optional[pulumi.Input[str]]):
+        pulumi.set(self, "ipv4_address", value)
+
+    @property
+    @pulumi.getter(name="macAddress")
+    def mac_address(self) -> Optional[pulumi.Input[str]]:
+        """
+        The mac address of the Linode.
+        """
+        return pulumi.get(self, "mac_address")
+
+    @mac_address.setter
+    def mac_address(self, value: Optional[pulumi.Input[str]]):
+        pulumi.set(self, "mac_address", value)
 
 
 @pulumi.input_type